requirements.txt
requests
beautifulsoup4

setuptools
python-telegram-bot==12.3.0
